fixed optimization flags in some Makefiles, tetgen wraper, launch medit, mosek wrapper
Former-commit-id: 68bbcbc66e895af55204bb60a38685be93abc43a

+++ include/igl/tetgen/tetgenio_to_tetmesh.cpp
@@ -53,7 +53,7 @@ IGL_INLINE bool igl::tetgenio_to_tetmesh(
   }
   assert(min_index >= 0);
   assert(max_index >= 0);
-  assert(max_index < V.size());
+  assert(max_index < (int)V.size());
 
   return true;
 }
